hugging face code/requirements.txt

yolov9pip==0.0.4
gradio==3.15.0
torch
huggingface-hub
